Move signal code into new bootstrap package
